+++ simulations/vip-492/abi/ResilientOracle.json
